fix MarkdownTokenParser on new changes
